backend/requirements.txt
Django>=4.2
djangorestframework
boto3
python-dotenv
requests
psycopg2-binary
